Added checks to ensure only the owner or admin can delete

+++ Projekt/web-platform/Controllers/SecurityIssuePostController.cs
@@ -169,6 +169,11 @@ namespace web_platform.Controllers
         [HttpPost]
         public async Task<ActionResult> DeleteSecurityIssuePost(int securityIssuePostId)
         {
+            var securityIssuePostToFind = await _ISecurityIssuePostService.GetById(securityIssuePostId);
+
+            // We need to ensure that any manual requests to this endpoint are still validated against the current user in case the UI gets bypassed
+            if (_userManager.GetUserId(User) != securityIssuePostToFind.ApplicationUser.Id || !User.IsInRole("Administrator")) { return Unauthorized(); }
+
             await _ISecurityIssuePostService.DeleteSecurityIssuePost(securityIssuePostId);
             return RedirectToAction("Index");
         }
